requirements.txt
streamlit
pandas
openpyxl
Pillow
numpy<2
scipy
torch
torchvision
transformers
timm
einops
accelerate